+++ changelog.md
@@ -4,5 +4,6 @@
 ### daris-core-services
 
 ### daris-portal:
+1. enabled AAF support in the login dialog.